Add s3 loader to new, refactored file structure
